fix: address Greptile review findings on email templates
- Move NotificationTemplateRow/INotificationTemplateRepository types to
  notification-repository-types.ts; original file re-exports for compat
- Use TRPCError NOT_FOUND in updateTemplate instead of plain Error
- Wrap previewTemplate Handlebars.compile calls in try-catch with BAD_REQUEST
- Add logger.info audit calls to updateTemplate and seedDefaults mutations
- Inject currentYear into HandlebarsRenderer render context
