+++ src/trpc/notification-template-router.ts
@@ -4,8 +4,10 @@
  * All procedures require platform_admin role via adminProcedure.
  */
 
+import { TRPCError } from "@trpc/server";
 import Handlebars from "handlebars";
 import { z } from "zod";
+import { logger } from "../config/logger.js";
 import { DEFAULT_TEMPLATES } from "../email/default-templates.js";
 import type { INotificationTemplateRepository } from "../email/notification-template-repository.js";
 import { adminProcedure, router } from "./init.js";
@@ -37,7 +39,7 @@ export function createNotificationTemplateRouter(getRepo: () => INotificationTem
         const repo = getRepo();
         const existing = await repo.getByName(input.name);
         if (!existing) {
-          throw new Error(`Template "${input.name}" not found`);
+          throw new TRPCError({ code: "NOT_FOUND", message: `Template "${input.name}" not found` });
         }
         await repo.upsert(input.name, {
           subject: input.subject ?? existing.subject,
@@ -45,6 +47,10 @@ export function createNotificationTemplateRouter(getRepo: () => INotificationTem
           textBody: input.textBody ?? existing.textBody,
           description: input.description ?? existing.description,
           active: input.active ?? existing.active,
+        });
+        logger.info("Notification template updated", {
+          action: "notification_template.update",
+          templateName: input.name,
         });
       }),
 
@@ -58,15 +64,27 @@ export function createNotificationTemplateRouter(getRepo: () => INotificationTem
         }),
       )
       .mutation(({ input }) => {
-        const subject = Handlebars.compile(input.subject)(input.data);
-        const html = Handlebars.compile(input.htmlBody)(input.data);
-        const text = Handlebars.compile(input.textBody)(input.data);
-        return { subject, html, text };
+        try {
+          const subject = Handlebars.compile(input.subject)(input.data);
+          const html = Handlebars.compile(input.htmlBody)(input.data);
+          const text = Handlebars.compile(input.textBody)(input.data);
+          return { subject, html, text };
+        } catch (err) {
+          throw new TRPCError({
+            code: "BAD_REQUEST",
+            message: `Template render error: ${err instanceof Error ? err.message : String(err)}`,
+          });
+        }
       }),
 
     seedDefaults: adminProcedure.mutation(async () => {
       const repo = getRepo();
       const inserted = await repo.seed(DEFAULT_TEMPLATES);
+      logger.info("Notification template defaults seeded", {
+        action: "notification_template.seed",
+        inserted,
+        total: DEFAULT_TEMPLATES.length,
+      });
       return { inserted, total: DEFAULT_TEMPLATES.length };
     }),
   });
